+++ tests/integration/devnet/test_srpc_gets.py
@@ -63,10 +63,10 @@ def test_get_gas_anyaddress_pass(sui_client: SuiClient):
     for anyaddress in addresses:
         inaddr = SuiAddress.from_hex_string(anyaddress)
         gas_objects = get_gas(sui_client, inaddr)
-        assert len(gas_objects) > 0
-        gas_balances = [gas.balance for gas in gas_objects]
-        total_balance = sum(gas_balances)
-        assert total_balance > 0
+        if gas_objects:
+            gas_balances = [gas.balance for gas in gas_objects]
+            total_balance = sum(gas_balances)
+            assert total_balance > 0
 
 
 def test_committee(sui_client: SuiClient) -> None: